Refactor imports for @hedvig-ui/Input
